fix(qlinear): quantize bias to int16 to avoid overflow
When quantizing a Linear with float16 weights, trying to quantize the bias
to int32 will always lead to an overflow during the clamp operation as the
integer boundaries of the projection interval cannot be represented in float16.
16-bit of precision is also more than sufficient to preserve accuracy even with
a non-optimal scale.

+++ quanto/quantization/nn/qlinear.py
@@ -31,9 +31,9 @@ class QLinear(QModuleMixin, torch.nn.Linear):
             if isinstance(qbias, QTensor):
                 if torch.any(qbias._scale != bias_scale):
                     # This should only happen if we calibrate again a frozen module
-                    qbias = qbias.rescale(torch.int32, bias_scale)
+                    qbias = qbias.rescale(torch.int16, bias_scale)
             else:
-                qbias = QTensor.quantize(qbias, torch.int32, bias_scale)
+                qbias = QTensor.quantize(qbias, torch.int16, bias_scale)
         return qweight, qbias
 
     def qweight(self):
